backend/requirements.txt
anthropic
fastapi
uvicorn
python-dotenv
pydantic
ruff